Miha: -Showing percentage for download -Changed opening of favourited items -Removing bookmarks

+++ fotona/FIFeaturedViewController.m
@@ -155,7 +155,7 @@
             }
             
             if (indexPath.section == count) {
-                newsSelected = indexPath.row;
+                newsSelected = (int)indexPath.row;
                 FIFeaturedNewsTableViewCell *cell = [self.tableViewFeatured cellForRowAtIndexPath:indexPath];
                 cell.signNewNewsCell.hidden = true;
                 [APP_DELEGATE setNewsArray:newsArray];
@@ -209,37 +209,23 @@
     }
     if (indexPath.row > newsCount - 1) {
         
-//        MBProgressHUD *hud=[[MBProgressHUD alloc] initWithView:self.view];
-//        [self.view addSubview:hud];
-//        hud.labelText = @"Loading...";
-//        //        [[MBProgressHUD showHUDAddedTo:self.view animated:YES]  setLabelText:@"Loading"];
-//        [hud show:YES];
-        
-        
+
         CGPoint offset = self.tableViewFeatured.contentOffset;
         offset.y-=10;
-//        self.tableViewFeatured.scrollEnabled =NO;
         [self.tableViewFeatured setContentOffset:offset animated:NO];
         int l = 4;
         if (newsCount + l > newsArray.count) {
-            l = newsArray.count - newsCount;
+            l = (int)newsArray.count - newsCount;
         }
         newsCount+=l;
         dispatch_queue_t queue = dispatch_queue_create("com.4egenus.fotona", NULL);
         dispatch_async(queue, ^{
-            newsArray = [FNews getImages:newsArray fromStart:indexPath.row forNumber:l];
+            newsArray = [FNews getImages:newsArray fromStart:(int)indexPath.row forNumber:l];
             dispatch_async(dispatch_get_main_queue(), ^{
-//                self.tableViewFeatured.scrollEnabled  = YES;
-//                [self.tableViewFeatured reloadData];
-//                [MBProgressHUD hideAllHUDsForView:self.view animated:YES];
-                
-                
                 NSMutableArray *rowsToReload =[[NSMutableArray alloc] init];
-                
                 for (int i = 0; i < l; i++){
                     NSIndexPath* index = [NSIndexPath indexPathForRow:indexPath.row+i inSection:indexPath.section];
                     [rowsToReload addObject:index];
-                    
                }
                [tableViewFeatured reloadRowsAtIndexPaths:rowsToReload withRowAnimation:UITableViewRowAnimationNone];
